Add to_event method to job class
Also split it into its own file
