+++ lambdas/scheduler/src/schedule/__init__.py
@@ -1,4 +1,5 @@
-from .hyp3_jobs import hyp3_jobs, Job
+from .hyp3_jobs import hyp3_jobs
+from .job import Job
 from . import queries
 
 __all__ = ['hyp3_jobs', 'queries', 'Job']
