+++ lambdas/scheduler/src/schedule/hyp3_jobs.py
@@ -1,42 +1,8 @@
-from typing import NamedTuple
-
 import hyp3_db
-from hyp3_db.hyp3_models import Subscription, User, Process
-from hyp3_events import NewGranuleEvent, StartEvent, EmailEvent
 from scheduler_env import environment
 
 from . import queries
-
-
-class Job(NamedTuple):
-    sub: Subscription
-    process: Process
-    user: User
-    granule: NewGranuleEvent
-
-    def is_notify_only(self):
-        print(self.process.name.lower())
-        return 'notify' in self.process.name.lower()
-
-    def to_start_event(self):
-        return StartEvent(
-            granule=self.granule.name,
-            user_id=self.user.id,
-            sub_id=self.sub.id,
-            output_patterns=self.process.output_patterns,
-            script_path=self.process.script,
-            additional_info=[]
-        )
-
-    def to_email_event(self):
-        return EmailEvent(
-            user_id=self.user.id,
-            sub_id=self.sub.id,
-            additional_info=[],
-            browse_url=self.granule.browse_url,
-            download_url=self.granule.download_url,
-            granule_name=self.granule.name
-        )
+from .job import Job
 
 
 def hyp3_jobs(new_granules):
